Implement more feedback
