+++ public/index.html
@@ -5,7 +5,7 @@
     <meta charset="utf-8">
     <meta content="IE=edge,chrome=1" http-equiv="X-UA-Compatible">
     <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-    <title>API Reference</title>
+    <title>API Documentation</title>
 
     <style>
       .highlight table td { padding: 5px; }
@@ -21,8 +21,6 @@
 }
 .highlight .gp {
   color: #555555;
-}
-.highlight .gs {
 }
 .highlight .gu {
   color: #aaaaaa;
@@ -611,7 +609,7 @@ Register and get a token to be inserted in the header</p>
             <span class="s2">"updatedAt"</span><span class="p">:</span> <span class="s2">"2017-08-21"</span>
         <span class="p">}</span>
     <span class="p">],</span>
-    <span class="s2">"paginationDetails"</span><span class="p">:</span> <span class="p">{</span>
+    <span class="s2">"pagination"</span><span class="p">:</span> <span class="p">{</span>
         <span class="s2">"page"</span><span class="p">:</span> <span class="mi">1</span><span class="p">,</span>
         <span class="s2">"pageCount"</span><span class="p">:</span> <span class="mi">1</span><span class="p">,</span>
         <span class="s2">"pageSize"</span><span class="p">:</span> <span class="mi">3</span><span class="p">,</span>
@@ -633,7 +631,7 @@ Register and get a token to be inserted in the header</p>
 <p>Response:</p>
 </blockquote>
 <pre class="highlight javascript tab-javascript"><code><span class="p">{</span>
-  <span class="p">[</span>
+  <span class="s2">"userDetails"</span><span class="p">:</span> <span class="p">[</span>
     <span class="p">{</span>
       <span class="s2">"id"</span><span class="p">:</span> <span class="mi">2</span><span class="p">,</span>
       <span class="s2">"username"</span><span class="p">:</span> <span class="s2">"seunadex"</span><span class="p">,</span>
@@ -741,7 +739,7 @@ Register and get a token to be inserted in the header</p>
             <span class="s2">"createdAt"</span><span class="p">:</span> <span class="s2">"2017-08-21"</span>
           <span class="p">}</span>
       <span class="p">],</span>
-      <span class="s2">"paginationDetails"</span><span class="p">:</span> <span class="p">{</span>
+      <span class="s2">"pagination"</span><span class="p">:</span> <span class="p">{</span>
           <span class="s2">"page"</span><span class="p">:</span> <span class="mi">1</span><span class="p">,</span>
           <span class="s2">"pageCount"</span><span class="p">:</span> <span class="mi">1</span><span class="p">,</span>
           <span class="s2">"pageSize"</span><span class="p">:</span> <span class="mi">3</span><span class="p">,</span>
@@ -783,7 +781,7 @@ Register and get a token to be inserted in the header</p>
         <span class="s2">"createdAt"</span><span class="p">:</span> <span class="s2">"2017-08-21"</span>
       <span class="p">}</span>
   <span class="p">],</span>
-  <span class="s2">"paginationDetails"</span><span class="p">:</span> <span class="p">{</span>
+  <span class="s2">"pagination"</span><span class="p">:</span> <span class="p">{</span>
     <span class="s2">"page"</span><span class="p">:</span> <span class="mi">1</span><span class="p">,</span>
     <span class="s2">"pageCount"</span><span class="p">:</span> <span class="mi">1</span><span class="p">,</span>
     <span class="s2">"pageSize"</span><span class="p">:</span> <span class="mi">2</span><span class="p">,</span>
