+++ server/routes/userRoute.js
@@ -1,4 +1,5 @@
 import UserControllers from '../controllers/UserControllers';
+import Authorization from '../middleware/Authorization';
 import Middleware from '../middleware/Middleware';
 import Validation from '../middleware/Validation';
 
@@ -8,25 +9,25 @@ export default (router) => {
   router.post('/api/v1/users/login', Validation.validateLoginInput,
   UserControllers.login);
 
-  router.get('/api/v1/users/:id', Middleware.verifyUser,
-  Middleware.findById,
+  router.get('/api/v1/users/:id', Authorization.verifyUser,
+  Middleware.findUser,
     UserControllers.show);
 
-  router.get('/api/v1/users', Middleware.verifyUser,
-  Middleware.verifyAdmin, UserControllers.index);
+  router.get('/api/v1/users', Authorization.verifyUser,
+  Authorization.verifyAdmin, UserControllers.index);
 
-  router.get('/api/v1/search/users', Middleware.verifyUser,
+  router.get('/api/v1/search/users', Authorization.verifyUser,
     UserControllers.search);
 
-  router.get('/api/v1/users/:id/documents', Middleware.verifyUser,
-  Middleware.findById,
+  router.get('/api/v1/users/:id/documents', Authorization.verifyUser,
+  Middleware.findUser,
     UserControllers.listUserDocuments);
 
-  router.put('/api/v1/users/:id', Middleware.verifyUser,
-  Middleware.verifyCurrentUser,
+  router.put('/api/v1/users/:id', Authorization.verifyUser,
+  Authorization.verifyCurrentUser,
     UserControllers.update);
 
-  router.delete('/api/v1/users/:id', Middleware.verifyUser,
-  Middleware.findById,
+  router.delete('/api/v1/users/:id', Authorization.verifyUser,
+  Middleware.findUser,
     UserControllers.destroy);
 };
